Error Code List from asset

+++ Arrowgene.Ddon.Shared/Csv/ClientErrorCodeCsvReader.cs
@@ -8,14 +8,14 @@ namespace Arrowgene.Ddon.Shared.Csv
 
         protected override ClientErrorCode CreateInstance(string[] properties)
         {
-            if (!int.TryParse(properties[0], out int messageId)) return null;
-            if (!int.TryParse(properties[1], out int errorId)) return null;
+            if (!uint.TryParse(properties[0], out uint messageId)) return null;
+            if (!uint.TryParse(properties[1], out uint errorId)) return null;
             string errorCode = properties[2];
             string msgJp = properties[3];
             string msgEn = properties[3];
             return new ClientErrorCode
             {
-                Id = errorId,
+                Id = (int)errorId,
                 MessageId = messageId,
                 ErrorCode = errorCode,
                 MessageJp = msgJp,
