+++ Arrowgene.Ddon.Shared/Model/ClientErrorCode.cs
@@ -3,8 +3,8 @@
     public class ClientErrorCode : IAsset
     {
         public int Id { get; set; }
-        public int ErrorId => Id;
-        public int MessageId { get; set; }
+        public uint ErrorId => (uint)Id;
+        public uint MessageId { get; set; }
         public string ErrorCode { get; set; }
         public string MessageJp { get; set; }
         public string MessageEn { get; set; }
